Move Tweet class to correct package.
