+++ mobile/src/main/java/jackSzm/androiddevtweets/TweetView.java
@@ -7,8 +7,6 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import net.jackszm.twittertagscanner.Tweet;
-
 public class TweetView extends LinearLayout {
 
     private ImageView authorAvatarImageView;
